Написал тесты

+++ src/main/java/ru/abramov/FinFlow/FinFlow/repository/CategoryRepository.java
@@ -10,7 +10,6 @@ import java.util.Optional;
 
 @Repository
 public interface CategoryRepository extends JpaRepository<Category, Long> {
-    Optional<Category> findById(long id);
-
+    Optional<Category> findById(int id);
 
 }
